pyproject.toml
[project]
name = "demo-se347"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]",
    "aiofiles",
    "mcp>=1.17.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
